refactor: Update references after refactoring
