+++ frontend/src/app/projects/project-detail/project-metadata/project-metadata.component.ts
@@ -3,7 +3,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-
 import {
   Component,
   EventEmitter,
@@ -13,7 +12,7 @@ import {
   SimpleChanges,
 } from '@angular/core';
 import { FormControl } from '@angular/forms';
-import { ToastService } from 'src/app/toast/toast.service';
+import { ToastService } from 'src/app/helpers/toast/toast.service';
 import {
   Project,
   ProjectService,
