+++ frontend/src/app/projects/project-overview/project-overview.component.ts
@@ -5,7 +5,7 @@
 
 import { Component, OnInit } from '@angular/core';
 import { Subscription, tap } from 'rxjs';
-import { NavBarService } from 'src/app/navbar/service/nav-bar.service';
+import { NavBarService } from 'src/app/general/navbar/service/nav-bar.service';
 import {
   ProjectService,
   UserMetadata,
