+++ frontend/src/app/sessions/active-sessions/upload-dialog/upload-dialog.component.ts
@@ -15,7 +15,7 @@ import {
 import { BehaviorSubject, Subscription } from 'rxjs';
 import { PathNode, Session } from 'src/app/schemes';
 import { LoadFilesService } from 'src/app/services/load-files/load-files.service';
-import { ToastService } from 'src/app/toast/toast.service';
+import { ToastService } from 'src/app/helpers/toast/toast.service';
 import { FileExistsDialogComponent } from './file-exists-dialog/file-exists-dialog.component';
 @Component({
   selector: 'upload-dialog',
