+++ frontend/src/app/sessions/session-overview/session-overview.component.ts
@@ -7,9 +7,9 @@ import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
 import { MatDialog } from '@angular/material/dialog';
 import { DeleteSessionDialogComponent } from '../delete-session-dialog/delete-session-dialog.component';
-import { NavBarService } from '../navbar/service/nav-bar.service';
-import { Session } from '../schemes';
-import { SessionService } from '../services/session/session.service';
+import { NavBarService } from '../../general/navbar/service/nav-bar.service';
+import { Session } from '../../schemes';
+import { SessionService } from '../../services/session/session.service';
 
 @Component({
   selector: 'app-session-overview',
