+++ frontend/src/app/settings/modelsources/t4c-settings/t4c-instance-settings/t4c-instance-settings.component.ts
@@ -3,7 +3,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-
 import { Component, OnInit, ViewChild } from '@angular/core';
 import {
   AbstractControl,
@@ -20,7 +19,7 @@ import {
 } from 'src/app/settings/modelsources/t4c-settings/service/t4c-repos/t4c-repo.service';
 import { T4CSyncService } from 'src/app/services/t4c-sync/t4-csync.service';
 import { ActivatedRoute } from '@angular/router';
-import { NavBarService } from 'src/app/navbar/service/nav-bar.service';
+import { NavBarService } from 'src/app/general/navbar/service/nav-bar.service';
 import { T4CRepoDeletionDialogComponent } from './t4c-repo-deletion-dialog/t4c-repo-deletion-dialog.component';
 
 @Component({
