+++ frontend/src/app/settings/modelsources/t4c-settings/t4c-settings.component.ts
@@ -3,9 +3,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-
 import { Component, OnInit } from '@angular/core';
-import { NavBarService } from 'src/app/navbar/service/nav-bar.service';
+import { NavBarService } from 'src/app/general/navbar/service/nav-bar.service';
 import { T4CSyncService } from 'src/app/services/t4c-sync/t4-csync.service';
 
 @Component({
