+++ frontend/src/app/settings/settings.component.html
@@ -21,22 +21,14 @@
 
 <div class="title">Modelsources</div>
 <div class="options-list">
-  <mat-card
-    *ngIf="integrations.git"
-    matRipple
-    [routerLink]="['/settings/modelsources/git']"
-  >
+  <mat-card matRipple [routerLink]="['/settings/modelsources/git']">
     <div class="subtitle">Git models</div>
     <img
       id="git"
       src="https://git-scm.com/images/logos/downloads/Git-Logo-1788C.png"
     />
   </mat-card>
-  <mat-card
-    *ngIf="integrations.t4c"
-    matRipple
-    [routerLink]="['/settings/modelsources/t4c']"
-  >
+  <mat-card matRipple [routerLink]="['/settings/modelsources/t4c']">
     <div class="subtitle">Team For Capella</div>
     <app-mat-icon size="50px">people</app-mat-icon>
   </mat-card>
